+++ RELEASE_NOTES.md
@@ -4,6 +4,7 @@
  * #8: Support for configurable coordinate display
  * #9: New command `Map Preview (with projection)`, allowing for data to be previewed with a declared projection (required for formats like WKT, that have no inferred geographic projection)
  * #10: Fix subsequent previews not working properly
+ * #11: Configuration option to allow preview content to be dumped out to an external HTML file (for debugging purposes)
 
 # 0.2.2